pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.169.0",
    "google-auth>=2.39.0",
    "google-auth-oauthlib>=1.2.2",
    "openai==1.57.0",
    "pandas==2.2.3",
    "pillow==11.0.0",
    "streamlit==1.40.2",
]
